requirements.txt
groq
tavily-python
mcp
python-dotenv
requests
certifi